feat(nova): added city and topic to user settings
