+++ packages/luna/src/locales/ar-eg.json
@@ -456,6 +456,11 @@
   "settings.edit.personal.email.placeholder": "قم بكتابة ايميلك",
   "settings.edit.personal.phone-number": "رقم الهاتف",
   "settings.edit.personal.phone-number.placeholder": "قم بكتابة رقم هاتفك",
+  "settings.edit.personal.city": "المدينة",
+  "settings.edit.personal.city.placeholder": "من أين أنت",
+  "settings.edit.personal.topics.title": "إضافة المفضلات",
+  "settings.edit.personal.topics": "قم باختيار المجالات التي تريد التحدث عنها",
+  "settings.edit.personal.topics.placeholder": "الرياضة، السفر، التكنولوجيا",
   "settings.edit.password.title": "تغيير كلمة السر",
   "settings.edit.password.current": "كلمة السر الحالية",
   "settings.edit.password.new": "كلمة السر الجديدة",
@@ -470,6 +475,7 @@
   "settings.notifications.coming-soon.description": "بإمكانك التحكم في الاشعارات بمجرد ان يتم تفعيل هذه الخاصية.",
   "settings.save": "حفظ التغيرات",
   "profile.update.error": "لديك مشكلة في تحديدث بياناتك، الرجاء المحاولة لاحقاً",
+  "profile.update.success": "تم تحديث بياناتك بنجاح",
   "lessons.button.rebook": "إعادة الحجز",
   "lessons.button.join": "دخول الحصة",
   "lessons.button.find-tutors": "تصفح المعلمين",
